% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{bin}
\alias{bin}
\title{bin : Transformation of a continuous data vector into a binary vector}
\usage{
bin(x, prop = 0.5)
}
\arguments{
\item{x}{: continuous vector}

\item{prop}{: probability of success on each trial}
}
\value{
binairy vector
}
\description{
bin : Transformation of a continuous data vector into a binary vector
}
