% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{colSD}
\alias{colSD}
\title{colSD : SD for each column of a matrix}
\usage{
colSD(matrix)
}
\arguments{
\item{matrix}{a matrix}
}
\value{
sd for each column
}
\description{
colSD : SD for each column of a matrix
}
