% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{colVar}
\alias{colVar}
\title{colVar : variance for each column of a matrix}
\usage{
colVar(matrix)
}
\arguments{
\item{matrix}{a matrix}
}
\value{
variance for each column
}
\description{
colVar : variance for each column of a matrix
}
