% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{cor.ld.pas}
\alias{cor.ld.pas}
\title{cor.ld.pas : Correlation calculation with nearest neighbor in mDNA matrix (for multiple pas)}
\usage{
cor.ld.pas(meth, annot, mulpas = 1:5)
}
\arguments{
\item{meth}{matrice of mDNA}

\item{annot}{annotation file of the mDNA matrix (Illuminina type)}

\item{pas}{1 by default, corresponds to the nearest neighbor. 2 for the second closest neighbor and etc ...}
}
\value{
a matrix, with a correlation vector, a genetic distance vector.
}
\description{
cor.ld.pas : Correlation calculation with nearest neighbor in mDNA matrix (for multiple pas)
}
\details{
This function makes it possible to calculate the correlation with the nearest neighbor (or the second or third, etc.). Correlation of pearson by default.
This function orders the mDNA matrix following chromosomes and chromosome
positions and uses the cor.pas() function to calculate the correlation for each chromosome.
Moreover the distance between each neighbor is calculated.
The difference with cor.ld() is that cor.ld.pas() can be used by varying the "pas" (between nearest neighbors).
}
