% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{cor.paired}
\alias{cor.paired}
\title{cor.paired : Correlation calculation with nearest neighbor}
\usage{
cor.paired(meth, pas = 1)
}
\arguments{
\item{meth}{matrice of mDNA (the matrix needs to be ordered according to the chromosomes and the chromosomal position)}

\item{pas}{1 by default, corresponds to the nearest neighbor. 2 for the second closest neighbor and etc ...}
}
\value{
a vector of correlation
}
\description{
cor.paired : Correlation calculation with nearest neighbor
}
\details{
This function makes it possible to calculate the correlation with the nearest neighbor (or the second or third, etc.). Correlation of pearson by default.
}
