% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{cor.pas}
\alias{cor.pas}
\title{cor.pas : Correlation calculation with nearest neighbor (with genetic distance)}
\usage{
cor.pas(meth, pas = 1)
}
\arguments{
\item{meth}{matrice of mDNA (the matrix needs to be ordered according to the chromosomes and the chromosomal position)}

\item{pas}{1 by default, corresponds to the nearest neighbor. 2 for the second closest neighbor and etc ...}
}
\value{
a vector of correlation
}
\description{
cor.pas : Correlation calculation with nearest neighbor (with genetic distance)
}
\details{
This function makes it possible to calculate the correlation with the nearest neighbor (or the second or third, etc.). Correlation of pearson by default.
The function cor.pas and different from the functions cor.paired, for not > 1 it does not calculate the correlation with all the neighbors (see the code of the function to understand).
}
