% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{F1.LD}
\alias{F1.LD}
\title{F1.LD : Calcul F1 score for a simulation (on effect sizes), take in account the linkage desequilibrum}
\usage{
F1.LD(beta, causal, nb.hit, neighbour.c = 50, neighbour.p = 50)
}
\arguments{
\item{beta}{: effect from a regression}

\item{causal}{: true causal variable (SNP or CpG)}

\item{nb.hit}{: number of first hit for calculate the F1 score}
}
\value{
the F1 score, the precision and the recall
}
\description{
F1.LD : Calcul F1 score for a simulation (on effect sizes), take in account the linkage desequilibrum
}
