% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{F1}
\alias{F1}
\title{F1 : Calcul F1 score for a simulation (on effect sizes)}
\usage{
F1(beta, causal, nb.hit = 10)
}
\arguments{
\item{beta}{: effect from a regression}

\item{causal}{: true causal variable (SNP or CpG)}

\item{nb.hit}{: number of first hit for calculate the F1 score}
}
\value{
the F1 score, the precision and the recall
}
\description{
F1 : Calcul F1 score for a simulation (on effect sizes)
}
