% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{F1.SCORE.FOR.BMA}
\alias{F1.SCORE.FOR.BMA}
\title{F1.SCORE.FOR.BMA : F1 score for bigmma package (for simulation)}
\usage{
F1.SCORE.FOR.BMA(bma, causal, M = NULL)
}
\arguments{
\item{bma}{result of the data.org.big function}

\item{causal}{true result of the simulation}

\item{M}{matrix of methylation}
}
\value{
precision : statistical precision of the test : TP / (TP + FP)

recall : statistical recall of the test : TP / (TP + FN)

f1_score : F1 Score of the test : 2(Pre * Rec) / (Pre + Rec)

length.list : number of pValues lower than "ral" (Bonferroni correction).

TP : True positive

FP : False positive

FN : False negative
}
\description{
F1.SCORE.FOR.BMA : F1 score for bigmma package (for simulation)
}
\details{
F1 = 2*recall*power/(recall + power)
}
