% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{F1.SCORE.FOR.SPARSE}
\alias{F1.SCORE.FOR.SPARSE}
\title{F1.SCORE.FOR.SPARSE : F1 score for sparse LFMM package (for simulation)}
\usage{
F1.SCORE.FOR.SPARSE(a, b, causal)
}
\arguments{
\item{a}{alpha effect in mediation (sparse_lfmm$B[,1])}

\item{b}{beta effect in mediation (sparse_lfmm$B[,2])}

\item{causal}{true result of the simulation}
}
\value{
precision : statistical precision of the test : TP / (TP + FP)

recall : statistical recall of the test : TP / (TP + FN)

f1_score : F1 Score of the test : 2(Pre * Rec) / (Pre + Rec)

length.list : number of pValues lower than "ral" (Bonferroni correction).

TP : True positive

FP : False positive

FN : False negative
}
\description{
F1.SCORE.FOR.SPARSE : F1 score for sparse LFMM package (for simulation)
}
\details{
Select the markers with a * b != 0
F1 = 2*recall*power/(recall + power)
}
