% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{f1.score.prs}
\alias{f1.score.prs}
\title{f1.score.prs : Calcul of the F1 score and the Polygenic Risk Score (PRS) for sparse regression}
\usage{
f1.score.prs(sparse.effect, causal, genotype, phenotype)
}
\arguments{
\item{sparse.effect}{: effect from a sparse regression}

\item{causal}{: true causal SNP}

\item{genotype}{: genotype --> one row per SNP, and one column per sample.}

\item{phenotype}{: phenotype of interest}
}
\value{
Return the F1 Score, the precision, the recall, the prediction of the phenotype and the r2 between the phenotype and x.pred.
}
\description{
f1.score.prs : Calcul of the F1 score and the Polygenic Risk Score (PRS) for sparse regression
}
