% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemma.R
\name{gemma}
\alias{gemma}
\title{gemma : GEMMA programme in R}
\usage{
gemma(geno, pheno, map, covar = NULL, gemma.exe = "./gemma.macosx",
  opt.mat = c(1, 2), met.reg = c("lmm", "bslmm"), lmm = c(1, 2, 3,
  4), bslmm = c(1, 2, 3))
}
\arguments{
\item{geno}{: genotype --> one row per SNP, and one column per sample.}

\item{pheno}{: phenotype d'interet (doit etre une matrice)}

\item{map}{: This file contains one line per SNP, with}

\item{covar}{: covariable et NULL marche}

\item{gemma.exe}{: the executable gemma}

\item{opt.mat}{: option for the calcul of the related.matrix, "1" calculates the centered relatedness # matrix while "2" calculates the standardized relatedness matrix}

\item{met.reg}{: linear mixed model of Bayesian sparse linear mixed model.}

\item{lmm}{: if lmm choose option : where the  "1" performs Wald test "2" performs likelihood ratio test, "3" performs score test, "4" performs all the three tests.}

\item{bslmm}{: if bslmm choose option : "1" fits a standard linear BSLMM "2" fits a ridge regression/GBLUP, "3" fits a probit BSLMM.}

\item{three}{columns: (1) SNP label, (2) base-pair position, (3) chromosome.}
}
\value{
: a gemma result
}
\description{
gemma : GEMMA programme in R
}
\details{
See paper of GEMMA
}
