% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{gemma2}
\alias{gemma2}
\title{gemma2 : GEMMA programme in R (better version of gemma)}
\usage{
gemma2(geno, pheno, map, covar = NULL, gemma.exe = "./gemma.macosx",
  opt.mat = c(1, 2), met.reg = c("lmm", "bslmm"), lmm = c(1, 2, 3,
  4), bslmm = c(1, 2, 3), pmin = 0.01, pmax = 0.05,
  opt.bs = "-w  10000 -s  10000")
}
\arguments{
\item{geno}{: genotype --> one row per SNP, and one column per sample.}

\item{pheno}{: phenotype d'interet (doit etre une matrice)}

\item{map}{: This file contains one line per SNP, with}

\item{covar}{: covariable et NULL marche}

\item{gemma.exe}{: the executable gemma}

\item{opt.mat}{: option for the calcul of the related.matrix, "1" calculates the centered relatedness # matrix while "2" calculates the standardized relatedness matrix}

\item{met.reg}{: linear mixed model of Bayesian sparse linear mixed model.}

\item{lmm}{: if lmm choose option : where the  "1" performs Wald test "2" performs likelihood ratio test, "3" performs score test, "4" performs all the three tests.}

\item{bslmm}{: if bslmm choose option : "1" fits a standard linear BSLMM "2" fits a ridge regression/GBLUP, "3" fits a probit BSLMM.}

\item{pmin}{: for bslmm : initial no zero proportion of beta (pi)}

\item{pmax}{: for bslmm : max no zero proportion of beta (pi)}

\item{opt.bs}{: option for bslmm : w and s default as in Zeng 2017 (10000 for both)}

\item{three}{columns: (1) SNP label, (2) base-pair position, (3) chromosome.}
}
\value{
: a gemma result
}
\description{
gemma2 : GEMMA programme in R (better version of gemma)
}
\details{
See paper of GEMMA
}
