% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{lfmm_lasso2}
\alias{lfmm_lasso2}
\title{lfmm_lasso2 : lfmm lasso with no scale on data}
\usage{
lfmm_lasso2(Y, X, K, nozero.prop = 0.01, lambda.num = 100,
  lambda.min.ratio = 0.01, lambda = NULL, it.max = 100,
  relative.err.epsilon = 1e-06)
}
\description{
lfmm_lasso2 : lfmm lasso with no scale on data
}
