% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{lfmm_ridge2}
\alias{lfmm_ridge2}
\title{lfmm_ridge2 : lfmm ridge with no scale on data}
\usage{
lfmm_ridge2(Y, X, K, lambda = 1e-05, algorithm = "analytical",
  it.max = 100, relative.err.min = 1e-06)
}
\description{
lfmm_ridge2 : lfmm ridge with no scale on data
}
