% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{med.test}
\alias{med.test}
\title{med.test : The Sobel mediation test}
\usage{
med.test(mv, iv, dv, cof = NULL)
}
\arguments{
\item{mv}{The mediator variable (aka M)}

\item{iv}{The independant variable (aka X)}

\item{dv}{The dependent variable (aka Y)}

\item{cof}{cofacteur}
}
\value{
pValue et score for Sobel test, Aroian test and Goodman test.
}
\description{
To compute statistics and p-values for the Sobel test. Results for
three versions of "Sobel test" are provided: Sobel test, Aroian test and Goodman test.
Function adapt from package "bda". (for covariable)
}
\details{
To test whether a mediator carries the influence on an IV to a DV.
Missing values are not allowed.
}
