% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{panel.cor}
\alias{panel.cor}
\title{panel.cor : pairs function}
\usage{
panel.cor(x, y, digits = 2, prefix = "", cex.cor, ...)
}
\arguments{
\item{x}{a data.frame}
}
\description{
panel.cor : pairs function
}
