% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{pwer.effect}
\alias{pwer.effect}
\title{pwer.effect : Calculate the power and the calibration of statiscal test (on the effects sizes)}
\usage{
pwer.effect(effect, causal, threshold = seq(0, 0.9, 0.05))
}
\arguments{
\item{effect}{effects sizes from a model}

\item{causal}{true result of the simulation}

\item{threshold}{the different thresholds where the power and fdr will be calculated}
}
\value{
the power and the fdr for each threshold
}
\description{
pwer.effect : Calculate the power and the calibration of statiscal test (on the effects sizes)
}
