% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{r_mediation_natural}
\alias{r_mediation_natural}
\title{r_mediation_natural : function to simulate DNA methylation data for mediation analyzis (based on NaturalGWAS model)}
\usage{
r_mediation_natural(meth = NULL, eff.size.x = 100, eff.size.y = 100,
  n.x = 10, n.y = 10, n.xy = 5, k = NULL, chr = NULL,
  pos = NULL, window = 101, pca = NULL)
}
\arguments{
\item{meth}{: methylation matrix}

\item{eff.size.x}{: effect size of X on M}

\item{n.x}{: number of causal CpGs associated with X}

\item{n.y}{: number of causal CpGs associated with y}

\item{n.xy}{: number of true mediators}

\item{k}{: number of latent factors (need to be estimated with a pca)}

\item{chr}{: chromosome}

\item{pos}{: chromosome position}

\item{window}{: distance between causal CpGs}

\item{pca}{: a former pca with svd() function, speed the simulation (NULL by default)}

\item{eff.size.x}{: effect size of Y on M}
}
\value{
causal.x : set of CpGs associated with the exposure

causal.y : set of CpGs associated with the outcome

causal.xy : set of true mediators

X : simulate phenotype

Y : simulate outcome
}
\description{
r_mediation_natural : function to simulate DNA methylation data for mediation analyzis (based on NaturalGWAS model)
}
\details{
Adaptation of the "simu_pheno" function of the "naturalgwas" package to simulate mediation data in high dimension
}
\examples{
# Simulate data :

}
