% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{rank.pwer}
\alias{rank.pwer}
\title{rank.pwer : Calculate the power and the calibration of statiscal test (for simulation)}
\usage{
rank.pwer(pval, known.mediator = NULL, toplist = 20, decreasing = F,
  ral = 0.05)
}
\arguments{
\item{pval}{pValues from a test}

\item{known.mediator}{true result of the simulation}

\item{toplist}{number of high pValues to use to calculate rank power}

\item{decreasing}{default F. This function can be use with effects size and with decreasing = T}

\item{ral}{default 0.05. Threshold for the calculation of the power and the recall}
}
\value{
cur : the rank power curve

auc : the AUC of the rank power curve

auc.max : the maximum AUC for a perfect model

auc.norm : auc/auc.max. Normalization of the AUC of the rank power curve

precision : statistical precision of the test : TP / (TP + FP)

recall : statistical recall of the test : TP / (TP + FN)

f1_score : F1 Score of the test : 2(Pre * Rec) / (Pre + Rec)

length.list : number of pValues lower than "ral" (Bonferroni correction).

TP : True positive

FP : False positive

FN : False negative
}
\description{
rank.pwer : Calculate the power and the calibration of statiscal test (for simulation)
}
\details{
AUC corresponds to the air under the curve (AUC) of a hits rank curve.
This curve is obtained by looking for hits in a toplist of the 20 CpGs with the lowest pValues.
F1 = 2*recall*power/(recall + power)
}
