% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{rank.pwer.sparse}
\alias{rank.pwer.sparse}
\title{rank.pwer.sparse : Calcul the AUC of the Hits Rank Curve for sparse regresion}
\usage{
rank.pwer.sparse(sparse.effect, causal)
}
\arguments{
\item{sparse.effect}{: effect from a sparse regression}

\item{causal}{: true causal SNP}
}
\value{
Return the AUC of the Hits Rank Curve
}
\description{
rank.pwer.sparse : Calcul the AUC of the Hits Rank Curve for sparse regresion
}
