% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{refactor2}
\alias{refactor2}
\title{refactor2 : fonction refactor}
\usage{
refactor2(data, k, covar = NULL, t = 500, numcomp = NULL,
  stdth = 0.02)
}
\arguments{
\item{data}{methylation matrix}

\item{k}{number of K}

\item{covar}{covariables}

\item{t}{number of of CpGs for the estimation of latent factor}

\item{numcomp}{like K}

\item{stdth}{sites were excluded due to low variance}
}
\description{
refactor2 : fonction refactor
}
\details{
see refactor paper
}
