% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{RMSE}
\alias{RMSE}
\title{RMSE : Calcul Root Square Mean Error}
\usage{
RMSE(beta, true.beta)
}
\arguments{
\item{beta}{: effect from a linear regression}

\item{true}{: true effect from a simulation}
}
\value{
the RMSE
}
\description{
RMSE : Calcul Root Square Mean Error
}
