% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{rowVar}
\alias{rowVar}
\title{rowVar : variance for each row of a matrix}
\usage{
rowVar(matrix)
}
\arguments{
\item{matrix}{a matrix}
}
\value{
variance for each row
}
\description{
rowVar : variance for each row of a matrix
}
