% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{sob.parallel}
\alias{sob.parallel}
\title{sob.parallel : The Sobel mediation test for high dimension}
\usage{
sob.parallel(X, Y, M, conf = NULL, nb.core = 2)
}
\arguments{
\item{X}{exposure}

\item{Y}{outcome}

\item{M}{methylation matrix}

\item{conf}{latent factors}

\item{nb.core}{number of core used (parallel)}
}
\value{
pValue et score for Sobel test, Aroian test and Goodman test.
}
\description{
To compute statistics and p-values for the Sobel test. Results for
three versions of "Sobel test" are provided: Sobel test, Aroian test and Goodman test.
Function adapt from package "bda". (for covariable)
}
\details{
To test whether a mediator carries the influence on an IV to a DV.
Missing values are not allowed.
}
