% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFonc.r
\name{write.gemma.covariates}
\alias{write.gemma.covariates}
\title{write.gemma.covariates}
\usage{
write.gemma.covariates(file, covariates, pheno)
}
\arguments{
\item{covariates}{covariable}
}
\value{
nothing
}
\description{
write.gemma.covariates
}
